requirements/common.txt
Django==1.5.1
South==0.8.1
Pinax==0.9a2
django-celery-with-redis==3.0
django-extensions==1.1.1
django-forms-bootstrap==2.0.3.post1
django-pipeline==1.2.21
django-user-accounts==1.0b8
gunicorn==0.17.2
numpy==1.7.1
pinax-theme-bootstrap-account==1.0b2
psycopg2==2.5.1
requests==1.2.3
